code cleanup inside Cart i2
